+++ app/build.gradle
@@ -102,7 +102,7 @@ dependencies {
     kapt "com.google.dagger:hilt-android-compiler:$hilt_version"
     implementation 'com.google.android.flexbox:flexbox:3.0.0'
 
-    /* Kotlin */
+    /* JetBrains */
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
 
     /* Other Java */
