+++ app/src/main/cpp/skyline/input/npad_device.h
@@ -6,12 +6,6 @@
 #include <kernel/types/KEvent.h>
 #include "shared_mem.h"
 
-namespace skyline::constant {
-    constexpr jlong MsInSecond{1000}; //!< The amount of milliseconds in a single second of time
-    constexpr jint AmplitudeMax{std::numeric_limits<u8>::max()}; //!< The maximum amplitude for Android Vibration APIs
-    constexpr i8 NullIndex{-1}; //!< The placeholder index value when there is no device present
-}
-
 namespace skyline::input {
     /**
      * @brief The orientations the Joy-Con(s) can be held in
@@ -137,8 +131,9 @@ namespace skyline::input {
 
       public:
         NpadId id;
-        i8 index{constant::NullIndex}; //!< The index of the device assigned to this player
-        i8 partnerIndex{constant::NullIndex}; //!< The index of a partner device, if present
+        static constexpr i8 NullIndex{-1}; //!< The placeholder index value when there is no device present
+        i8 index{NullIndex}; //!< The index of the device assigned to this player
+        i8 partnerIndex{NullIndex}; //!< The index of a partner device, if present
         NpadVibrationValue vibrationLeft; //!< Vibration for the left Joy-Con (Handheld/Pair), left LRA in a Pro-Controller or individual Joy-Cons
         std::optional<NpadVibrationValue> vibrationRight; //!< Vibration for the right Joy-Con (Handheld/Pair) or right LRA in a Pro-Controller
         NpadControllerType type{};
