+++ app/src/main/cpp/skyline/services/lm/ILogger.cpp
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: MPL-2.0
 // Copyright © 2020 Skyline Team and Contributors (https://github.com/skyline-emu/)
 
+#include <sstream>
 #include "ILogger.h"
 
 namespace skyline::service::lm {
